refactor(procedures): rename dm/ to device/ and nm/ to network/
More descriptive directory names for procedure families.
