+++ xknx/management/procedures/network/nm_individual_address_serial_number_write.py
@@ -6,7 +6,7 @@ import logging
 from typing import TYPE_CHECKING
 
 from xknx.exceptions import ManagementConnectionError
-from xknx.management.procedures.nm.nm_individual_address_serial_number_read import (
+from xknx.management.procedures.network.nm_individual_address_serial_number_read import (
     nm_individual_address_serial_number_read,
 )
 from xknx.telegram import apci
